requirements.txt
pandas==1.1.5
datetime==4.3
mysql-connector-python==8.0.23
mysqlclient==2.0.1
flask-sqlalchemy==2.5.1
flask-restful==0.3.8
flask==1.1.2
pyjwt==2.0.1
passlib==1.7.4
requests==2.25.1
flask-httpauth==4.4.0

